Species

Species within the Star Wars universe.

Usage

species

Format

A data frame with 37 rows and 10 variables:

name

The name of this species.

classification

The classification of this species.

designation

The designation of this species.

average_height

The average height of this person in centimeters.

skin_colors

A comma-seperated string of common skin colors for this species, none if this species does not typically have skin.

hair_colors

A comma-seperated string of common hair colors for this species, none if this species does not typically have hair.

eye_colors

A comma-seperated string of common eye colors for this species, none if this species does not typically have eyes.

average_lifespan

The average lifespan of this species in years.

homeworld

The URL of a planet resource, a planet that this species originates from.

language

The language commonly spoken by this species.

References

The Star Wars API (SWAPI).