Star Wars Data Schema

Includes information about the schema of the tables that were sourced from SWAPI, the Star Wars API. Not all properties returned from the API are columns in the data in this package: some properties were refactored into separate tables. For example, I combined the starships/ and vehicles/ endpoint into a single table. Both API endpoints returned a "pilots" property, which is described in the schema as an array of people who piloted the vehicle. The information in this property has been extracted into a separate table called pilots in the starwarsdb package.

Usage

schema

Format

A data frame with 5 rows and 4 variables:

endpoint

The name of the SWAPI endpoint.

endpoint_title

The title of the SWAPI endpoint.

endpoint_description

The description of the SWAPI endpoint.

properties

The properties of the endpoint as a nested table containing the variable, the data type, a description and the format of the property.

References

The Star Wars API (SWAPI).