Relational Data from the Star Wars API for Learning and Teaching

starwarsdb provides tables about the films, characters, planets and vehicles in the Star Wars movie franchise universe. The package and the tables have been organized to facilitate learning relational database concepts, like joins. It also includes functions that connect to a local DuckDB (file-based) database for teaching and practice working with remote databases via dplyr, DBI, and dm.

Tables

films

people

planets

species

vehicles

pilots

films_people

films_planets

films_vehicles

schema

References

The Star Wars API (SWAPI).

See also

Author

Maintainer: Garrick Aden-Buie garrick@adenbuie.com (ORCID)