Starships or Vehicles

A Starship or vehicle in the Star Wars universe.

Usage

vehicles

Format

A data frame with 75 rows and 14 variables:

name

The name of this vehicle. The common name, such as Sand Crawler.

type

The type of the vehicle: starship or vehicle.

class

The class of the vehicle, source from starship_class or vehicle_class.

model

The model or official name of this vehicle. Such as All Terrain Attack Transport.

manufacturer

The manufacturer of this vehicle. Comma seperated if more than one.

cost_in_credits

The cost of this vehicle new, in galactic credits.

length

The length of this vehicle in meters.

max_atmosphering_speed

The maximum speed of this vehicle in atmosphere.

crew

The number of personnel needed to run or pilot this vehicle.

passengers

The number of non-essential people this vehicle can transport.

cargo_capacity

The maximum number of kilograms that this vehicle can transport.

consumables

The maximum length of time that this vehicle can provide consumables for its entire crew without having to resupply.

hyperdrive_rating

The class of this starships hyperdrive.

MGLT

The Maximum number of Megalights this starship can travel in a standard hour. A Megalight is a standard unit of distance and has never been defined before within the Star Wars universe. This figure is only really useful for measuring the difference in speed of starships. We can assume it is similar to AU, the distance between our Sun (Sol) and Earth.

References

The Star Wars API (SWAPI).